Working algorithm version
Algorithm works and delivers all packages.  The final issue was the field I was using to pull the address for package 12.  Station was written "Station" in the package file, but the field I used in the distance table had "Sta".  The addresses would not match for loading and delivery... I changed the source file but did not update which field is used.